requirements.txt
yt-dlp==2022.9.1
python-vlc==3.0.16120